[Refactor] 타입 및 데이터 파일 정리
- EditorPanel에서 타입을 types 폴더로 이동
- 샘플 데이터 data 폴더로 이동
